Show locked pages from all sites (needs tests)
In order to let users easily switch between pages they currently work on across sites we show locked pages from all sites in the subnavigation tabs and on the dashboard widgets.

This also adds, that we show the site name and language code from the page in the page tab, to provide enough information for users to distinguish locked pages from each other.